requirements.txt
mcp>=0.1.0
anthropic>=0.25.0
google-generativeai>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0